# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayPass API
  description: |
    PayPass is a comprehensive payment platform providing secure, scalable payment processing solutions.
    
    ## Features
    - Payment Processing
    - Wallet Management  
    - User Authentication
    - Enterprise Billing
    - Fraud Detection
    - Compliance Management
    
    ## Authentication
    PayPass API uses API key authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```
    
    ## Rate Limiting
    API requests are rate limited based on your subscription plan:
    - Basic: 100 requests/hour
    - Premium: 1000 requests/hour  
    - Enterprise: 10000 requests/hour
    
    ## Webhooks
    PayPass can send webhook notifications for various events. Configure webhooks in your dashboard.
    
  version: 2.1.0
  contact:
    name: PayPass API Support
    email: api-support@paypass.com
    url: https://docs.paypass.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://paypass.com/terms
servers:
  - url: https://api.paypass.com/v1
    description: Production server
  - url: https://api-staging.paypass.com/v1
    description: Staging server
  - url: https://api-dev.paypass.com/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /wallets/{walletId}/balance:
    get:
      tags:
        - Wallets
      summary: Get wallet balance
      operationId: getWalletBalance
      parameters:
        - name: walletId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Balance'

  # Authentication Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Balance:
      type: object
      properties:
        available:
          type: number
          format: decimal
          description: Available balance
        pending:
          type: number
          format: decimal
          description: Pending balance
        currency:
          type: string
          description: Currency
        last_updated:
          type: string
          format: date-time
          description: Last update timestamp

    # Auth Schemas
...